requirements.txt
python-dotenv
streamlit==1.37.0
streamlit-webrtc
opencv-python
python-docx
python-pptx
pyPDF2
google-generativeai
moviepy
transformers
soundfile
torch
torchaudio
langchain
chromadb

# -e .